:recycle: Optimizing acess to data

+++ src/routes.ts
@@ -6,7 +6,14 @@ import Favorite from './models/Favorite';
 
 const router = express.Router();
 
-const getGameDetails = async (appId: string) => {
+interface ModelFavorite {
+  id: string;
+  appId: number;
+  token: string;
+  rating: number;
+}
+
+const getGameDetails = async (appId: number) => {
   try {
     let details = cache.get(appId);
     if (!details) {
@@ -22,7 +29,10 @@ const getGameDetails = async (appId: string) => {
     return undefined;
   }
 };
+
 const getFullGames = async () => {
+  console.log('get games API steam >>', new Date(Date.now()));
+
   try {
     let fullData = cache.get('fullData');
     if (!fullData) {
@@ -30,7 +40,8 @@ const getFullGames = async () => {
         'https://api.steampowered.com/ISteamApps/GetAppList/v0002/?format=json',
       );
       fullData = data.applist.apps;
-      cache.put('fullData', fullData);
+      const minute = 1000 * 60;
+      cache.put('fullData', fullData, minute * 120);
     }
 
     return fullData;
@@ -51,7 +62,7 @@ router.get('/', async (request, response) => {
 router.get('/:id', async (request, response, next) => {
   const appId = request.params.id;
   if (appId.match('\\d')) {
-    const gameDetails = await getGameDetails(appId);
+    const gameDetails = await getGameDetails(parseInt(appId, 10));
     if (gameDetails) {
       response.json(gameDetails);
     } else {
@@ -64,15 +75,21 @@ router.get('/:id', async (request, response, next) => {
 
 router.post('/favorite', async (request, response) => {
   try {
-    // const res = request.body;
     const repo = getRepository(Favorite);
 
     const model = {
       ...request.body,
       token: request.headers.token,
     };
-    const res = await repo.save(model);
-    return response.status(201).json(res);
+    let res = await repo.find({ where: model });
+    console.log(res.length);
+
+    if (res.length === 0) {
+      res = await repo.save(model);
+      console.log('if resSave');
+      return response.status(201).json(res);
+    }
+    return response.status(200).json(res);
   } catch (error) {
     console.log(error);
 
@@ -84,12 +101,20 @@ router.get('/favorite', async (request, response) => {
   try {
     const { token } = request.headers;
     const repo = getRepository(Favorite);
-    const res = await repo.find({ where: { token } });
+    const vecFavorites: ModelFavorite[] = await repo.find({ where: { token } });
+    if (vecFavorites.length === 0) {
+      return response.status(404).send();
+    }
+    const res = [];
+    for (let i = 0; i < vecFavorites.length; i++) {
+      res[i] = await getGameDetails(vecFavorites[i].appId);
+    }
     return response.status(200).json(res);
   } catch (error) {
-    return response.status(400).send();
+    return response.status(404).send();
   }
 });
+
 router.delete('/favorite/:appid', async (request, response) => {
   try {
     const appId = request.params.appid;
@@ -100,11 +125,14 @@ router.delete('/favorite/:appid', async (request, response) => {
     };
     const repo = getRepository(Favorite);
     const object = await repo.find({ where: model });
+    if (!object) {
+      return response.status(204);
+    }
     const res = await repo.remove(object);
-    return response.json(res);
+    return response.status(202).json(res);
   } catch (error) {
-    return response.status(400).send();
+    return response.status(404).send();
   }
 });
 
-export default router;
+export { router, getFullGames };
